check for admin rights

+++ src/at/lws/wnm/server/service/AuthorizationServiceImpl.java
@@ -7,6 +7,8 @@ import at.lws.wnm.server.dao.AuthorizationDao;
 import at.lws.wnm.server.dao.DaoRegistry;
 import at.lws.wnm.shared.model.Authorization;
 
+import com.google.appengine.api.users.UserService;
+import com.google.appengine.api.users.UserServiceFactory;
 import com.google.gwt.user.server.rpc.RemoteServiceServlet;
 
 public class AuthorizationServiceImpl extends RemoteServiceServlet implements
@@ -14,12 +16,13 @@ public class AuthorizationServiceImpl extends RemoteServiceServlet implements
 
 	private static final long serialVersionUID = 3902038151789960035L;
 	private final AuthorizationDao authorizationDao;
+	private UserService userService;
 
-	public AuthorizationServiceImpl()
-	{
+	public AuthorizationServiceImpl() {
 		authorizationDao = DaoRegistry.get(AuthorizationDao.class);
+		userService = UserServiceFactory.getUserService();
 	}
-	
+
 	@Override
 	public Collection<Authorization> queryAuthorizations() {
 
@@ -28,12 +31,28 @@ public class AuthorizationServiceImpl extends RemoteServiceServlet implements
 
 	@Override
 	public void storeAuthorization(Authorization aut) {
+		assertCurrentUserIsAdmin();
 		authorizationDao.storeAuthorization(aut);
 	}
 
 	@Override
 	public void deleteAuthorization(String email) {
+		assertCurrentUserIsAdmin();
 		authorizationDao.deleteAuthorization(email);
+	}
+
+	public void assertCurrentUserIsAdmin() {
+		if (!currentUserIsAdmin()) {
+			throw new IllegalStateException("current user is no admin: "
+					+ userService.getCurrentUser());
+		}
+	}
+
+	@Override
+	public boolean currentUserIsAdmin() {
+		final Authorization authorization = authorizationDao
+				.getAuthorization(userService.getCurrentUser());
+		return authorization != null && authorization.isAdmin();
 	}
 
 }
